Select and move some characters (#16)
* Changes

* Some refining

+++ src/Camera.cpp
@@ -10,7 +10,7 @@ Camera2D Camera_Init(int monitorWidth, int monitorHeight, int viewportWidth, int
 
     cam.zoom = zoom;
     // center the camera on the middle of the screen
-    cam.offset = (Vector2){monitorWidth / 2, monitorHeight / 2};
+    cam.offset = (Vector2){(float)monitorWidth / 2, (float)monitorHeight / 2};
 
     return cam;
 }
